Modifiche tempo reale e registrazione con controllo

+++ Pages/Registration.cshtml.cs
@@ -2,6 +2,9 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using FaceRecognition.Models;
 using FaceRecognition.Services;
+using System.Text.Json;
+using System.Text;
+
 
 namespace FaceRecognition.Pages
 {
@@ -10,7 +13,7 @@ namespace FaceRecognition.Pages
         public IFormFile? imageFile { get; set; }
         [TempData]
         public string? Message { get; set; }
-
+        public string? Error { get; set; }
         [BindProperty] 
         public FaceUser NewFace { get; set; } = default!;
         private readonly FaceService _service;
@@ -25,45 +28,46 @@ namespace FaceRecognition.Pages
         {
             if (imageFile != null && imageFile.Length > 0)
             {
-                NewFace.Id = GenerateRandomString(10);
-
-                string folderName = NewFace.Id;
-                string imageFileName = NewFace.Id + ".jpg";
-
-                string imageDirectory = Path.Combine("wwwroot", "imgs", folderName);
-
-                if (!Directory.Exists(imageDirectory))
+                if (ControllaImmagine(imageFile))
                 {
-                    Directory.CreateDirectory(imageDirectory);
+                    NewFace.Id = GenerateRandomString(10);
+
+                    string folderName = NewFace.Id;
+                    string imageFileName = NewFace.Id + ".jpg";
+
+                    string imageDirectory = Path.Combine("wwwroot", "imgs", folderName);
+
+                    if (!Directory.Exists(imageDirectory))
+                    {
+                        Directory.CreateDirectory(imageDirectory);
+                    }
+
+                    string imagePath = Path.Combine(imageDirectory, imageFileName);
+
+                    using (var memoryStream = new MemoryStream())
+                    {
+                        imageFile.CopyTo(memoryStream);
+                        byte[] imageBytes = memoryStream.ToArray();
+                        System.IO.File.WriteAllBytes(imagePath, imageBytes);
+                    }
+
+                    if (!ModelState.IsValid || NewFace == null)
+                        return Page();
+
+                    _service.AddFaceUsers(NewFace);
+
+                    Message = "Utente creato con successo e immagine salvata."; // Set the TempData message
+
+                    // Elimina il file 'representations_vgg_face.pkl' se esiste
+                    string pklFilePath = Path.Combine("wwwroot", "imgs",  "representations_vgg_face.pkl");
+                    if (System.IO.File.Exists(pklFilePath))
+                    {
+                        System.IO.File.Delete(pklFilePath);
+                    }
                 }
-
-                string imagePath = Path.Combine(imageDirectory, imageFileName);
-
-                using (var memoryStream = new MemoryStream())
-                {
-                    imageFile.CopyTo(memoryStream);
-                    byte[] imageBytes = memoryStream.ToArray();
-                    System.IO.File.WriteAllBytes(imagePath, imageBytes);
-                }
-
-                if (!ModelState.IsValid || NewFace == null)
-                    return Page();
-
-                _service.AddFaceUsers(NewFace);
-
-                 // Elimina il file 'representations_vgg_face.pkl' se esiste
-                string pklFilePath = Path.Combine("wwwroot", "imgs",  "representations_vgg_face.pkl");
-                if (System.IO.File.Exists(pklFilePath))
-                {
-                    System.IO.File.Delete(pklFilePath);
-                }
-
-                Message = "Utente creato con successo e immagine salvata."; // Set the TempData message
-
-                return RedirectToPage(); // Redirect to the same page
             }
 
-            return RedirectToPage("/Verify");
+            return Page();
         }
 
         public static string GenerateRandomString(int length)
@@ -88,5 +92,57 @@ namespace FaceRecognition.Pages
             return new string(result);
         }
 
+        public bool ControllaImmagine(IFormFile? image)
+        {
+            // Imposta il percorso di destinazione per salvare l'immagine 
+                string uploadsFolder = Path.Combine("wwwroot", "tmps");
+                string uniqueFileName = Guid.NewGuid().ToString() + "_webcam.jpg";
+                string filePath = Path.Combine(uploadsFolder, uniqueFileName);
+
+                // Salva l'immagine sul server
+                using (var stream = new FileStream(filePath, FileMode.Create))
+                {
+                   image.CopyTo(stream); 
+                }
+
+                using var httpClient = new HttpClient();
+
+                var json = new
+                {
+                    imagePath = filePath
+                };
+
+                var jsonContent = new StringContent(JsonSerializer.Serialize(json), Encoding.UTF8, "application/json");
+
+                var response = httpClient.PostAsync("http://127.0.0.1:5000/api/verifica", jsonContent).Result;
+
+                // Elimina il file temporaneo
+                if (System.IO.File.Exists(filePath))
+                {
+                    System.IO.File.Delete(filePath);
+                }
+
+                if (response.IsSuccessStatusCode)
+                {
+                    var result = response.Content.ReadAsStringAsync().Result;
+                   
+                    if (result == "0")
+                    {
+                        return true;
+                    }
+                    else if(result.Length == 10)
+                    {
+                        Error = "Errore, volto già registrato"; // Set the TempData message
+                        return false;
+                    }
+                    {
+                        Error = "Errore, impossibile identificare un volto"; // Set the TempData message
+                        return false;
+                    }
+                }
+
+            Error = "Errore, si prega di riprovare"; // Set the TempData message
+            return false;
+        }
     }
 }
